Update menu.es.toml
I would avoid "nosotros" which means us, because in Spanish you can also say "nosotras" which mens us but feminin. Currently many people discuss wether to visibilize women and gender diversities using "nosotres" (for men, women and gender diversities) should be used, but this is still not oficially accepted and not everyone likes that or uses that term. So I would just avoid using it and replacing it by "Sobre R-Ladies" which literaly means "about R-Ladies" intead of "about us".